move files to ts and add required types
